update: move examples from program.cs
